Extract component library code into seprate file
